+++ src/commonMain/kotlin/net/multigesture/kanama/api/Noise.kt
@@ -46,6 +46,16 @@ open class Noise(handle: MemorySegment) : Resource(handle) {
         return Image.wrap(ObjectCalls.ptrcallWithTwoIntTwoBoolDoubleBoolArgsRetObject(getSeamlessImageBind, handle, width, height, invert, in3dSpace, skirt, normalize))
     }
 
+    fun getImage3d(width: Int, height: Int, depth: Int, invert: Boolean = false, normalize: Boolean = true): List<Image> {
+        checkOpen()
+        return ObjectCalls.ptrcallWithThreeIntTwoBoolArgsRetTypedObjectList(getImage3dBind, handle, width, height, depth, invert, normalize, Image::fromHandle)
+    }
+
+    fun getSeamlessImage3d(width: Int, height: Int, depth: Int, invert: Boolean = false, skirt: Double = 0.1, normalize: Boolean = true): List<Image> {
+        checkOpen()
+        return ObjectCalls.ptrcallWithThreeIntBoolDoubleBoolArgsRetTypedObjectList(getSeamlessImage3dBind, handle, width, height, depth, invert, skirt, normalize, Image::fromHandle)
+    }
+
     companion object {
         @JvmStatic
         fun fromHandle(handle: MemorySegment): Noise? =
@@ -87,6 +97,16 @@ open class Noise(handle: MemorySegment) : Resource(handle) {
         private const val GET_SEAMLESS_IMAGE_HASH = 2770743602L
         private val getSeamlessImageBind by lazy {
             ObjectCalls.getMethodBind("Noise", "get_seamless_image", GET_SEAMLESS_IMAGE_HASH)
+        }
+
+        private const val GET_IMAGE_3D_HASH = 3977814329L
+        private val getImage3dBind by lazy {
+            ObjectCalls.getMethodBind("Noise", "get_image_3d", GET_IMAGE_3D_HASH)
+        }
+
+        private const val GET_SEAMLESS_IMAGE_3D_HASH = 451006340L
+        private val getSeamlessImage3dBind by lazy {
+            ObjectCalls.getMethodBind("Noise", "get_seamless_image_3d", GET_SEAMLESS_IMAGE_3D_HASH)
         }
     }
 }