+++ src/commonMain/kotlin/net/multigesture/kanama/api/RenderingServer.kt
@@ -740,6 +740,16 @@ object RenderingServer {
     }
 
     /**
+     * Returns 3D texture data as an array of `Image`s for the specified texture `RID`.
+     *
+     * Generated from Godot docs: RenderingServer.texture_3d_get
+     */
+    @JvmStatic
+    fun texture3dGet(texture: RID): List<Image> {
+        return ObjectCalls.ptrcallWithRIDArgRetTypedObjectList(texture3dGetBind, singleton, texture, Image::fromHandle)
+    }
+
+    /**
      * Calculates new MipMaps for the given Drawable `texture`.
      *
      * Generated from Godot docs: RenderingServer.texture_drawable_generate_mipmaps
@@ -5133,6 +5143,18 @@ object RenderingServer {
     }
 
     /**
+     * Bakes the material data of the Mesh passed in the `base` parameter with optional
+     * `material_overrides` to a set of `Image`s of size `image_size`. Returns an array of `Image`s
+     * containing material properties as specified in `BakeChannels`.
+     *
+     * Generated from Godot docs: RenderingServer.bake_render_uv2
+     */
+    @JvmStatic
+    fun bakeRenderUv2(base: RID, materialOverrides: List<RID>, imageSize: Vector2i): List<Image> {
+        return ObjectCalls.ptrcallWithRIDRIDListVector2iArgsRetTypedObjectList(bakeRenderUv2Bind, singleton, base, materialOverrides, imageSize, Image::fromHandle)
+    }
+
+    /**
      * Creates a canvas and returns the assigned `RID`. It can be accessed with the RID that is
      * returned. This RID will be used in all `canvas_*` RenderingServer functions. Once finished with
      * your RID, you will want to free the RID using the RenderingServer's `free_rid` method. Canvas
@@ -6772,6 +6794,11 @@ object RenderingServer {
     private const val TEXTURE_2D_LAYER_GET_HASH = 2705440895L
     private val texture2dLayerGetBind by lazy {
         ObjectCalls.getMethodBind("RenderingServer", "texture_2d_layer_get", TEXTURE_2D_LAYER_GET_HASH)
+    }
+
+    private const val TEXTURE_3D_GET_HASH = 2684255073L
+    private val texture3dGetBind by lazy {
+        ObjectCalls.getMethodBind("RenderingServer", "texture_3d_get", TEXTURE_3D_GET_HASH)
     }
 
     private const val TEXTURE_DRAWABLE_GENERATE_MIPMAPS_HASH = 2722037293L
@@ -8667,6 +8694,11 @@ object RenderingServer {
     private const val INSTANCES_CULL_CONVEX_HASH = 2488539944L
     private val instancesCullConvexBind by lazy {
         ObjectCalls.getMethodBind("RenderingServer", "instances_cull_convex", INSTANCES_CULL_CONVEX_HASH)
+    }
+
+    private const val BAKE_RENDER_UV2_HASH = 1904608558L
+    private val bakeRenderUv2Bind by lazy {
+        ObjectCalls.getMethodBind("RenderingServer", "bake_render_uv2", BAKE_RENDER_UV2_HASH)
     }
 
     private const val CANVAS_CREATE_HASH = 529393457L